requirements.txt
streamlit
streamlit-mic-recorder
pydub
torch
transformers
openai-whisper